+++ lib/message/parser/badges.spec.ts
@@ -22,6 +22,25 @@ describe("./message/parser/badges", function () {
       );
     });
 
+    it("should parse prediction badge & badge-info", function () {
+      assert.deepStrictEqual(
+        parseSingleBadge("predictions/blue-1"),
+        new TwitchBadge("predictions", "blue-1")
+      );
+      assert.deepStrictEqual(
+        parseSingleBadge("predictions/foo bar\\n baz"),
+        new TwitchBadge("predictions", "foo bar\\n baz")
+      );
+      assert.deepStrictEqual(
+        parseSingleBadge("predictions/<<<<<< HEAD[15A⸝asdf/test"),
+        new TwitchBadge("predictions", "<<<<<< HEAD[15A,asdf/test")
+      );
+      assert.deepStrictEqual(
+        parseSingleBadge("predictions/<<<<<< HEAD[15A⸝asdf⸝/test/"),
+        new TwitchBadge("predictions", "<<<<<< HEAD[15A,asdf,/test/")
+      );
+    });
+
     it("should preserve non-integer versions as-is", function () {
       assert.deepStrictEqual(
         parseSingleBadge("vip/1.0"),